+++ src/Programs/Starcounter.UsageTracker/API/Installer/StartCollectionHandler.POST.cs
@@ -35,7 +35,7 @@ namespace Starcounter.Applications.UsageTrackerApp.API.Installer {
                         Db.Transaction(() => {
 
                             string serial = data.downloadId;
-                            int previousInstallationNo = int.Parse(data.installationNo.ToString());
+                            Int64 previousInstallationNo = Int64.Parse(data.installationNo.ToString());
 
                             Installation installation = new Installation(serial, previousInstallationNo);
                             InstallerStart item = new InstallerStart(installation);
