+++ src/Programs/Starcounter.UsageTracker/API/Starcounter/UsageCollectionHandler.POST.cs
@@ -34,7 +34,7 @@ namespace Starcounter.Applications.UsageTrackerApp.API.Starcounter {
                         Db.Transaction(() => {
 
                             string serial = data.downloadId;
-                            int installationNo = int.Parse(data.installationNo.ToString());
+                            Int64 installationNo = Int64.Parse(data.installationNo.ToString());
 
                             Installation installation = StarcounterCollectionHandler.AssureInstallation(installationNo, serial);
                             StarcounterUsage item = new StarcounterUsage(installation);
